# apps/api/docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "snapdone Go API. 인증 credential은 JWT가 아닌 opaque 서버 세션 토큰.",
        "title": "snapdone API",
        "contact": {},
        "version": "0.1"
    },
    "basePath": "/",
    "paths": {
...
        "/v1/auth/capabilities": {
            "get": {
                "description": "서버에 설정 · 구현된 로그인 provider. 인증이 비활성이면 503이다.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "로그인 수단 목록",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httpserver.CapabilitiesResponse"
                        }
                    },
                    "503": {
                        "description": "인증 비활성 (provider_unavailable)",
                        "schema": {
                            "$ref": "#/definitions/httpserver.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/v1/auth/handoff/exchange": {
            "post": {
                "description": "web 서버가 verifier cookie와 코드로 앱 세션 아래 child web 세션을 받는다. 코드는 한 번만 쓸 수 있다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "handoff"
                ],
                "summary": "WebView 핸드오프 교환",
                "parameters": [
                    {
                        "description": "handoff 코드 · web verifier · next",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/httpserver.HandoffExchangeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httpserver.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "잘못된 요청 · 틀린 proof · 만료 · 재사용 · parent 취소 (invalid_callback)",
                        "schema": {
                            "$ref": "#/definitions/httpserver.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "내부 오류 (provider_unavailable)",
                        "schema": {
                            "$ref": "#/definitions/httpserver.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "인증 비활성 (provider_unavailable)",
                        "schema": {
                            "$ref": "#/definitions/httpserver.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/auth/handoff/start": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "앱이 자기 Bearer로 WebView용 30초 일회용 코드를 받는다. root mobile 세션만 받는다.\nchallenge는 web 서버가 가진 verifier의 S256이고, next는 허용 목록(/ · /history) 안이어야 한다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "handoff"
                ],
                "summary": "WebView 핸드오프 코드 발급",
                "parameters": [
                    {
                        "description": "web verifier의 challenge · 이동할 web 경로",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/httpserver.HandoffStartRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httpserver.HandoffStartResponse"
                        }
                    },
                    "400": {
                        "description": "잘못된 요청 · challenge · next (invalid_callback)",
                        "schema": {
                            "$ref": "#/definitions/httpserver.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "credential 없음 · 무효 · root mobile 세션 아님 (session_expired)",
                        "schema": {
                            "$ref": "#/definitions/httpserver.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "내부 오류 (provider_unavailable)",
                        "schema": {
                            "$ref": "#/definitions/httpserver.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "인증 비활성 (provider_unavailable)",
                        "schema": {
                            "$ref": "#/definitions/httpserver.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/auth/logout": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "이 credential의 세션만 취소한다(root면 child 포함). 모르는 · 만료된 credential도 204다.",
                "tags": [
                    "auth"
                ],
                "summary": "로그아웃",
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Authorization 헤더 형식 오류 (session_expired)",
                        "schema": {
                            "$ref": "#/definitions/httpserver.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "내부 오류 (provider_unavailable)",
                        "schema": {
                            "$ref": "#/definitions/httpserver.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "인증 비활성 (provider_unavailable)",
                        "schema": {
                            "$ref": "#/definitions/httpserver.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/v1/auth/session": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Bearer credential의 세션을 확인하고 idle 만료를 연장한다. 만료 · 취소 · 모르는 credential은 401이다.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "현재 세션",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httpserver.SessionResponse"
                        }
                    },
                    "401": {
                        "description": "credential 없음 · 만료 · 취소 (session_expired)",
                        "schema": {
                            "$ref": "#/definitions/httpserver.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "내부 오류 (provider_unavailable)",
                        "schema": {
                            "$ref": "#/definitions/httpserver.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "인증 비활성 (provider_unavailable)",
                        "schema": {
                            "$ref": "#/definitions/httpserver.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "httpserver.CapabilitiesResponse": {
            "type": "object",
            "properties": {
                "providers": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "google"
                    ]
                }
            }
        },
        "httpserver.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "enum": [
                        "session_expired",
                        "provider_unavailable",
                        "invalid_callback",
                        "invalid_image",
                        "image_too_large",
                        "unsupported_image",
                        "job_not_found",
                        "invalid_onboarding",
                        "onboarding_complete",
                        "onboarding_out_of_order"
                    ],
                    "example": "session_expired"
                }
            }
        },
...
        "httpserver.HandoffExchangeRequest": {
            "type": "object",
            "required": [
                "code",
                "next",
                "verifier"
            ],
            "properties": {
                "code": {
                    "type": "string",
                    "example": "one-time-handoff-code"
                },
                "next": {
                    "type": "string",
                    "enum": [
                        "/",
                        "/history"
                    ],
                    "example": "/history"
                },
                "verifier": {
                    "type": "string",
                    "example": "base64url-web-verifier-43-to-128-chars"
                }
            }
        },
        "httpserver.HandoffStartRequest": {
            "type": "object",
            "required": [
                "challenge",
                "next"
            ],
            "properties": {
                "challenge": {
                    "type": "string",
                    "example": "base64url-S256-challenge-of-the-web-verifier"
                },
                "next": {
                    "type": "string",
                    "enum": [
                        "/",
                        "/history"
                    ],
                    "example": "/history"
                }
            }
        },
        "httpserver.HandoffStartResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "one-time-handoff-code"
                }
            }
        },
...
        "httpserver.LoginResponse": {
            "type": "object",
            "properties": {
                "credential": {
                    "type": "string",
                    "example": "opaque-session-credential"
                },
                "session": {
                    "$ref": "#/definitions/httpserver.SessionResponse"
                }
            }
        },
...
        "httpserver.SessionResponse": {
            "type": "object",
            "properties": {
                "expiresAt": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2026-10-01T00:00:00Z"
                },
                "onboardingStep": {
                    "type": "string",
                    "enum": [
                        "intro",
                        "purpose",
                        "first-image",
                        "complete"
                    ],
                    "example": "intro"
                },
                "user": {
                    "$ref": "#/definitions/httpserver.UserResponse"
                }
            }
        },
        "httpserver.UserResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "4f1c2a9e-0000-4000-8000-000000000000"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "\"Bearer \u003copaque session credential\u003e\" 형식. 서버만 해석할 수 있는 세션 토큰.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}